refactor(review): extract snapshot builder
